requirements.txt
pytest
selenium
allure-pytest
Appium-Python-Client